requirements.txt
langchain
langgraph
chromadb
pymongo
python-dotenv
langchain-community
ollama
langchain-ollama